Enhance the marker names
They are now much more stable.

+++ Sources/swift-sh/OptionGroups/BuildAndRunOptions.swift
@@ -27,14 +27,14 @@ final class BuildAndRunOptions : ParsableArguments {
 	var scriptOptions: ScriptOptions
 	
 	/* Returns the arguments that should be given to swift to run the script. */
-	func prepareRun(forceCopySource: Bool = false, logger: Logger) async throws -> (swiftArgs: [String], swiftFileToRun: String, swiftStdin: Data?, packageFolderPath: FilePath?, cleanup: () -> Void) {
+	func prepareRun(forBuilding: Bool = false, logger: Logger) async throws -> (swiftArgs: [String], swiftFileToRun: String, swiftStdin: Data?, packageFolderPath: FilePath?, cleanup: () -> Void) {
 		let fm = FileManager.default
-		let xdgDirs = try BaseDirectories(prefixAll: "swift-sh")
+		let xdgDirs = try BaseDirectories.swiftSH.get()
 		
 		let scriptSource = try (
 			!scriptPathIsContent ?
 				(
-					forceCopySource ?
+					forBuilding ?
 						ScriptSource(copying: FilePath(scriptPathOrContent), fileManager: fm) :
 						ScriptSource(path:    FilePath(scriptPathOrContent), fileManager: fm)
 				) :
@@ -52,21 +52,25 @@ final class BuildAndRunOptions : ParsableArguments {
 		}
 		
 		let scriptPathForSwift: String
-		var scriptData: (Data, hash: Data)?
+		var scriptData: Data?
+		var scriptHash: Data?
 		if let scriptPath = scriptSource.scriptPath {
 			scriptData = nil
+			scriptHash = scriptPathIsContent ? Data() : nil
 			scriptPathForSwift = scriptPath.0.string
 		} else {
-			scriptData = (Data(), hash: Data())
+			scriptData = Data()
+			scriptHash = Data()
 			scriptPathForSwift = "-"
 		}
 		
 		let depsPackage = try DepsPackage(
-			scriptSource: scriptSource, scriptData: &scriptData,
+			scriptSource: scriptSource, scriptData: &scriptData, scriptHash: &scriptHash,
 			useSSHForGithubDependencies: useSSHForGithubDependencies,
 			skipPackageOnNoRemoteModules: skipPackageOnNoRemoteModules,
 			fileManager: fm, logger: logger
 		)
+		
 		let (swiftArgs, packageFolderPath): ([String], FilePath?) = try await {
 			guard let depsPackage else {
 				return ([String](), nil)
@@ -74,7 +78,7 @@ final class BuildAndRunOptions : ParsableArguments {
 			/* Retrieve the REPL invocation in package folder path.
 			 * Note: This is not protected with regard to multiple scripts trying to use the same store entry.
 			 * TODO: Make the REPL invocation retrieval concurrent-safe. */
-			let packageFolderRelativePath = FilePath("store").appending(depsPackage.packageHash.reduce("", { $0 + String(format: "%02x", $1) }))
+			let packageFolderRelativePath = try xdgDirs.storeFolderRelativePath(for: depsPackage.packageHash)
 			let packageFolderPath = try xdgDirs.ensureCacheDirPath(packageFolderRelativePath)
 			let ret = try await depsPackage.retrieveREPLInvocation(
 				packageFolder: packageFolderPath,
@@ -82,25 +86,28 @@ final class BuildAndRunOptions : ParsableArguments {
 				disableSandboxForPackageResolution: disableSandboxForPackageResolution,
 				fileManager: fm, logger: logger
 			)
-			/* If retrieving the REPL invocation was successful, we mark the store entry as being used.
-			 * This is only for clients and has no use for swift-sh itself (allows light cleaning of the cache folder).
-			 * Note: We are not concurrent-safe for this either. */
-			let packageFolderAliasDiscriminator = Insecure.MD5
-				.hash(data: scriptData?.hash ?? Data(scriptPathForSwift.utf8))
-				.reduce("", { $0 + String(format: "%02x", $1) })
-			let markersFolderPath = try xdgDirs.ensureCacheDirPath("markers")
-			let packageFolderAliasPath = markersFolderPath.appending("\(scriptSource.scriptName)--\(packageFolderAliasDiscriminator)")
-			do {
-				try? fm.removeItem(at: packageFolderAliasPath.url)
-				/* Do not use the URL variant of this method as it is not possible (AFAICT) to create a relative link with it. */
-				try fm.createSymbolicLink(atPath: packageFolderAliasPath.string, withDestinationPath: FilePath("..").appending(packageFolderRelativePath.components).string)
-			} catch {
-				logger.warning("Failed to create marker link in swift-sh cache.", metadata: ["error": "\(error)", "marker-path": "\(packageFolderAliasPath.string)"])
+			
+			if !forBuilding {
+				/* If retrieving the REPL invocation was successful and we’re not building, we mark the store entry as being used.
+				 * This is only for clients and has no use for swift-sh itself (allows light cleaning of the cache folder).
+				 * Note: We are not concurrent-safe for this either. */
+				do {
+					let packageFolderAliasPath = try xdgDirs.ensureParentsForMarkerPathWith(
+						absoluteScriptPath: scriptSource.initialAbsoluteScriptPath,
+						nonReplayableScriptDataHash: scriptHash,
+						hashFunction: Insecure.MD5.self
+					)
+					try? fm.removeItem(at: packageFolderAliasPath.url)
+					/* Do not use the URL variant of this method as it is not possible (AFAICT) to create a relative link with it. */
+					try fm.createSymbolicLink(atPath: packageFolderAliasPath.string, withDestinationPath: FilePath("..").appending(packageFolderRelativePath.components).string)
+				} catch {
+					logger.warning("Failed to create marker link in swift-sh cache.", metadata: ["error": "\(error)"])
+				}
 			}
 			return (ret, packageFolderPath)
 		}()
 		
-		return (swiftArgs, scriptPathForSwift, scriptData?.0, packageFolderPath, cleanup)
+		return (swiftArgs, scriptPathForSwift, scriptData, packageFolderPath, cleanup)
 	}
 	
 }
